Remove useless abstract Simple* repositories
