+++ app/src/main/java/org/tokend/template/data/repository/AccountRepository.kt
@@ -5,13 +5,13 @@ import io.reactivex.Single
 import org.tokend.rx.extensions.toSingle
 import org.tokend.sdk.api.v3.accounts.params.AccountParamsV3
 import org.tokend.template.data.model.AccountRecord
-import org.tokend.template.data.repository.base.SimpleSingleItemRepository
+import org.tokend.template.data.repository.base.SingleItemRepository
 import org.tokend.template.di.providers.ApiProvider
 import org.tokend.template.di.providers.WalletInfoProvider
 
 class AccountRepository(private val apiProvider: ApiProvider,
                         private val walletInfoProvider: WalletInfoProvider)
-    : SimpleSingleItemRepository<AccountRecord>() {
+    : SingleItemRepository<AccountRecord>() {
 
     override fun getItem(): Observable<AccountRecord> {
         return getAccountResponse().toObservable()
