+++ app/src/main/java/org/tokend/template/data/repository/FeesRepository.kt
@@ -3,7 +3,7 @@ package org.tokend.template.data.repository
 import io.reactivex.Observable
 import io.reactivex.Single
 import org.tokend.rx.extensions.toSingle
-import org.tokend.template.data.repository.base.SimpleSingleItemRepository
+import org.tokend.template.data.repository.base.SingleItemRepository
 import org.tokend.template.di.providers.ApiProvider
 import org.tokend.template.di.providers.WalletInfoProvider
 import org.tokend.template.features.fees.model.FeeRecord
@@ -11,7 +11,7 @@ import org.tokend.template.features.fees.model.FeesRecords
 
 class FeesRepository(private val apiProvider: ApiProvider,
                      private val walletInfoProvider: WalletInfoProvider
-) : SimpleSingleItemRepository<FeesRecords>() {
+) : SingleItemRepository<FeesRecords>() {
     override fun getItem(): Observable<FeesRecords> {
         return getFeesResponse().toObservable()
     }
