+++ app/src/main/java/org/tokend/template/data/repository/KeyValueEntriesRepository.kt
@@ -7,14 +7,14 @@ import org.tokend.sdk.api.base.params.PagingParamsV2
 import org.tokend.sdk.utils.SimplePagedResourceLoader
 import org.tokend.template.data.model.KeyValueEntryRecord
 import org.tokend.template.data.repository.base.RepositoryCache
-import org.tokend.template.data.repository.base.SimpleMultipleItemsRepository
+import org.tokend.template.data.repository.base.MultipleItemsRepository
 import org.tokend.template.di.providers.ApiProvider
 import org.tokend.template.extensions.mapSuccessful
 
 class KeyValueEntriesRepository(
         private val apiProvider: ApiProvider,
         itemsCache: RepositoryCache<KeyValueEntryRecord>
-) : SimpleMultipleItemsRepository<KeyValueEntryRecord>(itemsCache) {
+) : MultipleItemsRepository<KeyValueEntryRecord>(itemsCache) {
     private val entriesMap = mutableMapOf<String, KeyValueEntryRecord>()
 
     override fun getItems(): Single<List<KeyValueEntryRecord>> {
