+++ app/src/main/java/org/tokend/template/data/repository/TfaFactorsRepository.kt
@@ -6,7 +6,7 @@ import org.tokend.rx.extensions.toCompletable
 import org.tokend.rx.extensions.toSingle
 import org.tokend.sdk.api.tfa.model.TfaFactor
 import org.tokend.template.data.repository.base.RepositoryCache
-import org.tokend.template.data.repository.base.SimpleMultipleItemsRepository
+import org.tokend.template.data.repository.base.MultipleItemsRepository
 import org.tokend.template.di.providers.ApiProvider
 import org.tokend.template.di.providers.WalletInfoProvider
 import org.tokend.template.features.tfa.model.TfaFactorCreationResult
@@ -16,7 +16,7 @@ class TfaFactorsRepository(
         private val apiProvider: ApiProvider,
         private val walletInfoProvider: WalletInfoProvider,
         itemsCache: RepositoryCache<TfaFactorRecord>
-) : SimpleMultipleItemsRepository<TfaFactorRecord>(itemsCache) {
+) : MultipleItemsRepository<TfaFactorRecord>(itemsCache) {
 
     override fun getItems(): Single<List<TfaFactorRecord>> {
         val signedApi = apiProvider.getSignedApi()
