+++ app/src/main/java/org/tokend/template/data/repository/assets/AssetsRepository.kt
@@ -8,7 +8,7 @@ import org.tokend.sdk.api.base.params.PagingParamsV2
 import org.tokend.sdk.api.v3.assets.params.AssetsPageParams
 import org.tokend.sdk.utils.SimplePagedResourceLoader
 import org.tokend.template.data.repository.base.RepositoryCache
-import org.tokend.template.data.repository.base.SimpleMultipleItemsRepository
+import org.tokend.template.data.repository.base.MultipleItemsRepository
 import org.tokend.template.di.providers.ApiProvider
 import org.tokend.template.di.providers.UrlConfigProvider
 import org.tokend.template.extensions.mapSuccessful
@@ -19,7 +19,7 @@ class AssetsRepository(
         private val urlConfigProvider: UrlConfigProvider,
         private val mapper: ObjectMapper,
         itemsCache: RepositoryCache<AssetRecord>
-) : SimpleMultipleItemsRepository<AssetRecord>(itemsCache) {
+) : MultipleItemsRepository<AssetRecord>(itemsCache) {
     override fun getItems(): Single<List<AssetRecord>> {
 
         val loader = SimplePagedResourceLoader(
