+++ app/src/main/java/org/tokend/template/data/repository/base/SingleItemRepository.kt
@@ -1,12 +1,19 @@
 package org.tokend.template.data.repository.base
 
+import io.reactivex.Completable
 import io.reactivex.Observable
+import io.reactivex.disposables.Disposable
+import io.reactivex.rxkotlin.subscribeBy
 import io.reactivex.subjects.BehaviorSubject
+import io.reactivex.subjects.CompletableSubject
 
 /**
  * Repository that holds a single [T] item.
  */
-abstract class SingleItemRepository<T> : Repository() {
+abstract class SingleItemRepository<T: Any> : Repository() {
+    private var updateResultSubject: CompletableSubject? = null
+    private var updateDisposable: Disposable? = null
+
     /**
      * Repository item
      */
@@ -22,7 +29,7 @@ abstract class SingleItemRepository<T> : Repository() {
         item?.let { itemSubject.onNext(it) }
     }
 
-    abstract protected fun getItem(): Observable<T>
+    protected abstract fun getItem(): Observable<T>
 
     protected open fun getStoredItem(): Observable<T> {
         return Observable.empty()
@@ -37,5 +44,56 @@ abstract class SingleItemRepository<T> : Repository() {
         item = newItem
 
         broadcast()
+    }
+
+    override fun update(): Completable {
+        invalidate()
+
+        return synchronized(this) {
+            val resultSubject = updateResultSubject.let {
+                if (it == null) {
+                    val new = CompletableSubject.create()
+                    updateResultSubject = new
+                    new
+                } else {
+                    it
+                }
+            }
+
+            isLoading = true
+
+            val storedItemObservable =
+                    if (isNeverUpdated) getStoredItem() else Observable.empty()
+
+            updateDisposable?.dispose()
+            updateDisposable = storedItemObservable.concatWith(
+                    getItem()
+                            .map {
+                                storeItem(it)
+                                it
+                            }
+            )
+                    .subscribeBy(
+                            onNext = { newItem: T ->
+                                onNewItem(newItem)
+                            },
+                            onComplete = {
+                                isNeverUpdated = false
+                                isLoading = false
+
+                                updateResultSubject = null
+                                resultSubject.onComplete()
+                            },
+                            onError = {
+                                isLoading = false
+                                errorsSubject.onNext(it)
+
+                                updateResultSubject = null
+                                resultSubject.onError(it)
+                            }
+                    )
+
+            resultSubject
+        }
     }
 }
