+++ app/src/main/java/org/tokend/template/features/kyc/storage/KycStateRepository.kt
@@ -17,7 +17,7 @@ import org.tokend.sdk.api.v3.requests.params.ChangeRoleRequestPageParams
 import org.tokend.sdk.api.v3.requests.params.RequestParamsV3
 import org.tokend.sdk.utils.extentions.getTypedRequestDetails
 import org.tokend.template.data.repository.BlobsRepository
-import org.tokend.template.data.repository.base.SimpleSingleItemRepository
+import org.tokend.template.data.repository.base.SingleItemRepository
 import org.tokend.template.di.providers.ApiProvider
 import org.tokend.template.di.providers.WalletInfoProvider
 import org.tokend.template.features.kyc.model.KycForm
@@ -31,7 +31,7 @@ class KycStateRepository(
         private val walletInfoProvider: WalletInfoProvider,
         private val submittedStatePersistor: SubmittedKycStatePersistor?,
         private val blobsRepository: BlobsRepository
-) : SimpleSingleItemRepository<KycState>() {
+) : SingleItemRepository<KycState>() {
     private class NoRequestFoundException : Exception()
 
     private data class KycRequestAttributes(
