+++ app/src/main/java/org/tokend/template/features/send/recipient/contacts/repository/ContactsRepository.kt
@@ -5,12 +5,12 @@ import android.provider.ContactsContract
 import io.reactivex.Single
 import io.reactivex.schedulers.Schedulers
 import org.tokend.template.data.repository.base.RepositoryCache
-import org.tokend.template.data.repository.base.SimpleMultipleItemsRepository
+import org.tokend.template.data.repository.base.MultipleItemsRepository
 import org.tokend.template.features.send.recipient.contacts.model.ContactRecord
 
 class ContactsRepository(val context: Context,
                          itemsCache: RepositoryCache<ContactRecord>
-) : SimpleMultipleItemsRepository<ContactRecord>(itemsCache) {
+) : MultipleItemsRepository<ContactRecord>(itemsCache) {
 
     override fun getItems(): Single<List<ContactRecord>> {
         val contacts = mutableListOf<ContactRecord>()
