+++ app/src/main/java/org/tokend/template/features/trade/orderbook/repository/OrderBookRepository.kt
@@ -3,7 +3,7 @@ package org.tokend.template.features.trade.orderbook.repository
 import io.reactivex.Observable
 import org.tokend.rx.extensions.toSingle
 import org.tokend.sdk.api.v3.orderbook.params.OrderBookParamsV3
-import org.tokend.template.data.repository.base.SimpleSingleItemRepository
+import org.tokend.template.data.repository.base.SingleItemRepository
 import org.tokend.template.di.providers.ApiProvider
 import org.tokend.template.features.trade.orderbook.model.OrderBook
 
@@ -12,7 +12,7 @@ class OrderBookRepository
         private val apiProvider: ApiProvider,
         private val baseAsset: String,
         private val quoteAsset: String
-) : SimpleSingleItemRepository<OrderBook>() {
+) : SingleItemRepository<OrderBook>() {
     override fun getItem(): Observable<OrderBook> {
         val api = apiProvider.getApi()
